Moved ShellManagerService to infra
