+++ Cairo Desktop/Cairo Desktop/Cairo Desktop.csproj
@@ -122,7 +122,6 @@
     <Compile Include="SupportingClasses\IWindowService.cs" />
     <Compile Include="SupportingClasses\MenuBarWindowService.cs" />
     <Compile Include="SupportingClasses\PicSumWallpaperClient.cs" />
-    <Compile Include="SupportingClasses\ShellManagerService.cs" />
     <Compile Include="SupportingClasses\TaskbarWindowService.cs" />
     <Compile Include="SupportingClasses\WindowManager.cs" />
     <Compile Include="SupportingClasses\FilePathValidationRule.cs" />
