+++ Cairo Desktop/Cairo Desktop/CairoSettingsWindow.xaml.cs
@@ -18,6 +18,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Interop;
+using CairoDesktop.Infrastructure.Services;
 using ManagedShell.Common.Helpers;
 using ManagedShell.Common.Logging;
 using OpenFileDialog = System.Windows.Forms.OpenFileDialog;
