+++ Cairo Desktop/Cairo Desktop/SupportingClasses/WindowManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Forms;
 using System.Windows.Threading;
+using CairoDesktop.Infrastructure.Services;
 using ManagedShell.AppBar;
 using ManagedShell.Common.Helpers;
 using Microsoft.Extensions.Logging;
